functions/requirements.txt
azure-functions==1.20.0
azure-mgmt-compute==31.0.0
azure-identity==1.17.1
azure-storage-queue==12.10.0
azure-mgmt-managementgroups==1.0.0
azure-mgmt-subscription==3.1.1
azure-mgmt-resourcegraph==8.0.0
azure-mgmt-monitor==6.0.2
azure-mgmt-alertsmanagement==2.0.0b2
azure-mgmt-network==26.0.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==22.0.0
requests==2.32.3
urllib3==2.2.1
pydantic==2.8.2
tenacity==9.0.0
croniter
pytz
holidays
datetime